+++ src/ft_memset.c
@@ -1,8 +1,8 @@
 #include <string.h>
 
-void			*ft_memset(void *b, int c, int len)
+void			*ft_memset(void *b, int c, size_t len)
 {
-	int				i;
+	size_t			i;
 	char			*tmp;
 
 	i = 0;
